Refactor date extension functions
