+++ agenda/data/src/main/java/com/nibalk/tasky/agenda/data/local/source/RoomLocalReminderDataSource.kt
@@ -7,14 +7,12 @@ import com.nibalk.tasky.agenda.data.local.mapper.toReminderEntity
 import com.nibalk.tasky.agenda.domain.model.AgendaItem
 import com.nibalk.tasky.agenda.domain.source.local.LocalReminderDataSource
 import com.nibalk.tasky.agenda.domain.source.local.ReminderId
-import com.nibalk.tasky.core.data.utils.toEndOfDayMillis
-import com.nibalk.tasky.core.data.utils.toStartOfDayMillis
 import com.nibalk.tasky.core.domain.util.DataError
 import com.nibalk.tasky.core.domain.util.Result
+import com.nibalk.tasky.core.domain.util.toEndOfDayMillis
+import com.nibalk.tasky.core.domain.util.toStartOfDayMillis
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
-import timber.log.Timber
 import java.time.LocalDate
 
 class RoomLocalReminderDataSource(
@@ -24,28 +22,19 @@ class RoomLocalReminderDataSource(
     override suspend fun getAllReminders(): Flow<List<AgendaItem.Reminder>> {
         return reminderDao.getAllReminders().map { entities ->
             entities.map { entity ->
-                Timber.d("[OfflineFirst-GetAll] LOCAL | reminder entity = %s", entity)
                 entity.toAgendaItemReminder()
             }
         }
     }
 
     override suspend fun getRemindersByDate(selectedDate: LocalDate): Flow<List<AgendaItem.Reminder>> {
-        Timber.d("[OfflineFirst-GetAll] LOCAL | selectedDate = %s", selectedDate)
-        Timber.d("[OfflineFirst-GetAll] LOCAL | toStartOfDayMillis = %s", selectedDate.toStartOfDayMillis())
-        Timber.d("[OfflineFirst-GetAll] LOCAL | toEndOfDayMillis = %s", selectedDate.toEndOfDayMillis())
-
-        val result = reminderDao.getAllRemindersByDate(
+        return reminderDao.getAllRemindersByDate(
             selectedDate.toStartOfDayMillis(), selectedDate.toEndOfDayMillis()
         ).map { entities ->
-            Timber.d("[OfflineFirst-GetAll] LOCAL | entities = %s", entities[0])
             entities.map { entity ->
-                Timber.d("[OfflineFirst-GetAll] LOCAL | entity = %s", entity)
                 entity.toAgendaItemReminder()
             }
         }
-        Timber.d("[OfflineFirst-GetAll] LOCAL | result = %s", result.toString())
-        return result
     }
 
     override suspend fun getReminderById(reminderId: String): AgendaItem.Reminder? {
