+++ agenda/data/src/main/java/com/nibalk/tasky/agenda/data/remote/mapper/AttendeeRemoteMapper.kt
@@ -2,8 +2,8 @@ package com.nibalk.tasky.agenda.data.remote.mapper
 
 import com.nibalk.tasky.agenda.data.remote.dto.EventAttendeeDto
 import com.nibalk.tasky.agenda.domain.model.EventAttendee
-import com.nibalk.tasky.core.data.utils.toLocalDateTime
-import com.nibalk.tasky.core.data.utils.toLongDate
+import com.nibalk.tasky.core.domain.util.toLocalDateTime
+import com.nibalk.tasky.core.domain.util.toLongDateTime
 
 fun EventAttendee.toEventAttendeeDto(): EventAttendeeDto {
     return EventAttendeeDto(
@@ -11,7 +11,7 @@ fun EventAttendee.toEventAttendeeDto(): EventAttendeeDto {
         email = email,
         fullName = fullName,
         isGoing = isGoing,
-        remindAt = remindAt.toLongDate()
+        remindAt = remindAt.toLongDateTime()
     )
 }
